Вернул тип Long при передаче значения declarationId в отчеты

+++ service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DeclarationDataServiceImpl.java
@@ -588,7 +588,7 @@ public class DeclarationDataServiceImpl implements DeclarationDataService {
                     zipXmlIn.getNextEntry();
                     Map<String, Object> params = new HashMap<String, Object>();
 
-                    params.put("declarationId", declarationData.getId().intValue());
+                    params.put("declarationId", declarationData.getId());
                     return createJasperReport(zipXmlIn, declarationTemplateService.getJrxml(declarationData.getDeclarationTemplateId()), jrSwapFile, params);
                 } catch (IOException e) {
                     throw new ServiceException(e.getLocalizedMessage(), e);
